+++ CHANGELOG.md
@@ -57,6 +57,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - make tags editable on attacks and ranges
 - show tags on equipment tab
 - equipped items can now be edited directly from the equipment page
+- modifiable initiative
+  ([#383](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/383))
 
 ### Changed